+++ ninja-core/src/main/java/ninja/exceptions/InternalServerErrorException.java
@@ -1,0 +1,84 @@
+/*
+ * Copyright 2014 ra.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package ninja.exceptions;
+
+/**
+ * A convenience unchecked exception. 
+ * Allows you to wrap any exception (checked or unchecked) and throw it.
+ * 
+ * Should signal a html error 500 (something went wrong on the server).
+ * 
+ * Useful inside controllers or filters for instance.
+ * 
+ * Ninja is supposed to pick it up and render an appropriate error page.
+ * 
+ */
+public class InternalServerErrorException extends RuntimeException {  
+    
+    /** 
+     * Constructs a new internal server error  exception with {@code null} as its
+     * detail message.  The cause is not initialized, and may subsequently be
+     * initialized by a call to {@link #initCause}.
+     */
+    public InternalServerErrorException() {
+        super();
+    }
+
+    /** 
+     * Constructs a new internal server error  exception with the specified detail message.
+     * The cause is not initialized, and may subsequently be initialized by a
+     * call to {@link #initCause}.
+     *
+     * @param   message   the detail message. The detail message is saved for
+     *          later retrieval by the {@link #getMessage()} method.
+     */
+    public InternalServerErrorException(String message) {
+        super(message);
+    }
+
+    /**
+     * Constructs a new internal server error  exception with the specified detail message and
+     * cause.  <p>Note that the detail message associated with
+     * {@code cause} is <i>not</i> automatically incorporated in
+     * this internal server error exception's detail message.
+     *
+     * @param  message the detail message (which is saved for later retrieval
+     *         by the {@link #getMessage()} method).
+     * @param  cause the cause (which is saved for later retrieval by the
+     *         {@link #getCause()} method).  (A <tt>null</tt> value is
+     *         permitted, and indicates that the cause is nonexistent or
+     *         unknown.
+     */
+    public InternalServerErrorException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    /** Constructs a new internal server error exception with the specified cause and a
+     * detail message of <tt>(cause==null ? null : cause.toString())</tt>
+     * (which typically contains the class and detail message of
+     * <tt>cause</tt>).  This constructor is useful for internal server error exceptions
+     * that are little more than wrappers for other throwables.
+     *
+     * @param  cause the cause (which is saved for later retrieval by the
+     *         {@link #getCause()} method).  (A <tt>null</tt> value is
+     *         permitted, and indicates that the cause is nonexistent or
+     *         unknown.)
+     */
+    public InternalServerErrorException(Throwable cause) {
+        super(cause);
+    }
+}