+++ ninja-core/src/main/java/ninja/utils/ResultHandler.java
@@ -27,15 +27,12 @@ import ninja.AsyncResult;
 import ninja.Context;
 import ninja.Renderable;
 import ninja.Result;
+import ninja.exceptions.InternalServerErrorException;
 import ninja.template.TemplateEngine;
 import ninja.template.TemplateEngineManager;
 import org.slf4j.Logger;
 
-/**
- * Handles the result
- * 
- * @author James Roper
- */
+
 @Singleton
 public class ResultHandler {
 
@@ -93,11 +90,9 @@ public class ResultHandler {
     private void handleRenderable(Renderable renderable,
                                   Context context,
                                   Result result) {
-        try {
-            renderable.render(context, result);
-        } catch (Exception e) {
-            logger.error("Error while handling renderable", e);
-        }
+        
+        renderable.render(context, result);
+
     }
 
     private void renderWithTemplateEngineOrRaw(Context context, Result result) {
@@ -127,8 +122,8 @@ public class ResultHandler {
                     
                     writer.write((String) result.getRenderable());
 
-                } catch (IOException e) {
-                    throw new RuntimeException(e);
+                } catch (IOException ioException) {
+                    throw new InternalServerErrorException(ioException);
                 }
 
             } else if (result.getRenderable() instanceof byte[]) {
@@ -146,15 +141,15 @@ public class ResultHandler {
 
                     outputStream.write((byte[]) result.getRenderable());
                     
-                } catch (IOException e) {
-                    throw new RuntimeException(e);
+                } catch (IOException ioException) {
+                    throw new InternalServerErrorException(ioException);
                 }
                 
             } else {
                 
                 context.finalizeHeaders(result);
                 
-                throw new IllegalArgumentException(
+                throw new InternalServerErrorException(
                         "No template engine found for result content type "
                                 + result.getContentType());
             }